requirements.txt
Flask==2.2.3
torch==2.0.0
torchvision==0.15.1
nltk==3.7
numpy==1.24.3
psycopg2-binary==2.9.5
SpeechRecognition==3.9.0
pydub==0.25.1
